Add log settings parameters to cmd
